split out modal for editing printer settings so can be changed with disabled
