+++ server/views/partials/PrinterManager/printerSettings.modal.ejs
@@ -57,74 +57,6 @@
                     </button>
                 </div>
             </div>
-            <h5>OctoPrint Host</h5>
-            <hr>
-            <p>Set your OctoPrint Host connection details</p>
-            <div id="psPrinterConnection">
-                <form>
-                    <div class="row pb-1">
-                        <div class="col-12 col-md-6">
-                            <label for="psPrinterName">Printer Name:</label>
-                            <div class="input-group">
-                                <input id="psPrinterName" type="text" class="form-control"
-                                       placeholder="OctoPrint User">
-                                <div id="generateNameButtonWrapper" class="input-group-append">
-
-                                </div>
-                            </div>
-                            <small class="form-text text-muted">Your current Printer Name. Use the generate button or update manually to create a new one</small>
-
-                            <small>Note: <code>This is not saved back to OctoPrint!</code></small>
-                        </div>
-                        <div class="col-12 col-md-6">
-                            <label for="psOctoPrintUser">OctoPrint User:</label>
-                            <select id="psOctoPrintUser" class="custom-select">
-
-                            </select>
-                            <small class="form-text text-muted">Current OctoPrint User that
-                                OctoFarm is using to connect with.</small>
-                        </div>
-                    </div>
-                    <div class="row pb-1">
-                        <div class="col-12 col-md-6">
-                            <label for="psPrinterURL">Printer URL:</label>
-                            <input id="psPrinterURL" type="text" class="form-control"
-                                   placeholder="Printer URL">
-                            <small class="form-text text-muted">URL of OctoPrint Host inc. port.
-                                Defaults to "http://" if not specified.</small>
-                            <small>Example: <code>http://192.168.1.5:81</code></small>
-                        </div>
-                        <div class="col-12 col-md-6">
-                            <label for="psWebSocketProtocol">Printer WebSocket Protocol:</label>
-                            <select id="psWebSocketProtocol" class="custom-select" disabled>
-                                <option value="ws://">ws://</option>
-                                <option value="wss://">wss://</option>
-                            </select>
-                            <small class="form-text text-muted">Websocket connection protocol. Automatically selected based on http/https value in Printer URL. (disabled)</small>
-                            <small>Example: <code>"ws:// (http) " or "wss:// (https)" </code></small>
-                        </div>
-                    </div>
-                    <div class="row pb-2">
-                        <div class="col-12 col-md-6">
-                            <label for="psAPIKEY">API Key:</label>
-                            <input id="psAPIKEY" type="text" class="form-control"
-                                   placeholder="API Key">
-                            <small class="form-text text-muted">OctoPrints API Key. It's
-                                recquired to use the User/Application API Key for OctoPrint
-                                version 1.4.1+.</small>
-                        </div>
-                        <div class="col-12 col-md-6">
-                            <label for="psCamURL">Camera URL:</label>
-                            <input id="psCamURL" type="text" class="form-control"
-                                   placeholder="Camera URL">
-                            <small class="form-text text-muted">URL of mjpeg camera stream.
-                                Defaults to "http://" if not specified.</small>
-                            <small>Example:
-                                <code>http://192.168.1.5/webcam/?action=stream</code></small>
-                        </div>
-                    </div>
-                </form>
-            </div>
             <h5>Serial Defaults</h5>
             <hr>
             <p id="portNotAvailableMessage"></p>
